Create packet.cpp file and move functions
